Comments; Remove some print statements; bug fix

+++ main.py
@@ -14,12 +14,13 @@ agents = {}
 seenMatches = []
 choice = ''
 
-print(os.listdir())
-if "Agent Yoinker U.exe" in os.listdir():
-    if "Agent Yoinker.exe" in os.listdir():
+#Update system by https://github.com/Neefs
+#Dont do comments often so bear with me
+if "Agent Yoinker U.exe" in os.listdir(): #checks if updated version is in dir. Removes old version and renames updated to main.
+    if "Agent Yoinker.exe" in os.listdir(): 
         os.remove("Agent Yoinker.exe")
-    os.rename("Agent Yoinker U.exe", "Agent Yoinker.exe")
-    print("in")
+    os.rename("Agent Yoinker U.exe", "Agent Yoinker.exe") 
+
 
 with open('data.json', 'r') as f:
     data = json.load(f)
@@ -36,10 +37,9 @@ with open('data.json', 'r') as f:
 if update == True:
     try:
         updatedata = requests.get("https://api.github.com/repos/deadly/valorant-agent-yoinker/releases/latest").json()
-        latest = float(updatedata["tag_name"].split("-")[1])
-    except Exception as e: #requests.ConnectionError
-        raise e
-        print("An error occured while checking for updates.")
+        latest = float(updatedata["tag_name"].split("-")[1]) #this is accounting for tagnames to be formatted like "something-1.1"
+    except requests.ConnectionError:
+        print("Could not connect to github server. Make sure you are connected to the internet.")
     else:
         if VERSION != latest:
             uchoice = input("There is a newer version would you like to update? (y or n): ").lower()
@@ -49,20 +49,23 @@ if update == True:
                     f.write(requests.get(updatedata["assets"][0]["browser_download_url"]).content)
                 
                 with ZipFile("temp.zip", "r") as zfo:
-                    zfo.extractall("./t")
+                    zfo.extractall("./t") #extracts zip
                 os.remove("temp.zip")
                 base = "t/"+os.listdir("t")[0]
                 with open("data.json", "wb") as wf:
                     with open(f"./{base}/data.json", "rb") as rf:
                         wf.write(rf.read())
+                        #updates config
                 
                 with open("Agent Yoinker U.exe", "wb") as wf:
                     with open(f"./{base}/Agent Yoinker.exe", "rb") as rf:
                         wf.write(rf.read())
-                shutil.rmtree("t")
+                        #updates main file
+                        #U is there because cant edit a file you are currently using
+                shutil.rmtree("t") # recurrsivly removes a dir
 
                 os.startfile("Agent Yoinker U.exe")
-                os._exit(1)
+                os._exit(1) # exit() got a nonetype error after building exe so using os._exit
